mod: Sorted Set으로 일간사용자 추적 및 매시간 삭제 스케줄링 추가

+++ src/main/java/OneQ/OnSurvey/domain/survey/model/dto/GlobalStats.java
@@ -4,13 +4,13 @@ public record GlobalStats (
     Long totalDueCount,
     Long totalCompletedCount,
     Long totalPromotionCount,
-    Integer dailyUserCount
+    Long dailyUserCount
 ){
     public static GlobalStats of(
         Long totalDueCount,
         Long totalCompletedCount,
         Long totalPromotionCount,
-        Integer dailyUserCount
+        Long dailyUserCount
     ) {
         return new GlobalStats(
             totalDueCount,
